+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The following emojis are used to highlight certain changes:
 
 ### Security
 
+- fix panic when incoming Bitswap protobuf message does not contain `Wantlist` [#961](https://github.com/ipfs/boxo/pull/961)
 
 ## [v0.32.0]